# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OpdVisitCancelRequest": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "OpdVisitCancelRequest",
        "type": "object"
      },
      "OpdVisitCompleteRequest": {
        "properties": {
          "create_billing_draft": {
            "default": true,
            "title": "Create Billing Draft",
            "type": "boolean"
          },
          "finalize_notes": {
            "default": true,
            "title": "Finalize Notes",
            "type": "boolean"
          },
          "follow_up_notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Notes"
          }
        },
        "title": "OpdVisitCompleteRequest",
        "type": "object"
      },
      "OpdVisitCreateRequest": {
        "properties": {
          "add_to_queue": {
            "default": true,
            "title": "Add To Queue",
            "type": "boolean"
          },
          "appointment_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Appointment Id"
          },
          "chief_complaint": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chief Complaint"
          },
          "doctor_id": {
            "format": "uuid",
            "title": "Doctor Id",
            "type": "string"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "patient_id": {
            "format": "uuid",
            "title": "Patient Id",
            "type": "string"
          },
          "queue_priority": {
            "default": "normal",
            "enum": [
              "normal",
              "urgent"
            ],
            "title": "Queue Priority",
            "type": "string"
          },
          "visit_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visit Date"
          },
          "visit_type": {
            "default": "walk_in",
            "enum": [
              "walk_in",
              "appointment"
            ],
            "title": "Visit Type",
            "type": "string"
          }
        },
        "required": [
          "patient_id",
          "doctor_id"
        ],
        "title": "OpdVisitCreateRequest",
        "type": "object"
      },
      "OpdVisitStartRequest": {
        "properties": {
          "doctor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Id"
          }
        },
        "title": "OpdVisitStartRequest",
        "type": "object"
      },
      "OpdVisitUpdateRequest": {
        "properties": {
          "chief_complaint": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chief Complaint"
          },
          "doctor_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doctor Id"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          }
        },
        "title": "OpdVisitUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Hospital Management SaaS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/opd/visits": {
      "get": {
        "description": "List OPD visits with filters.",
        "operationId": "list_visits_api_v1_opd_visits_get",
        "parameters": [
          {
            "in": "query",
            "name": "doctor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Doctor Id"
            }
          },
          {
            "in": "query",
            "name": "patient_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Patient Id"
            }
          },
          {
            "in": "query",
            "name": "visit_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Visit Date"
            }
          },
          {
            "in": "query",
            "name": "from_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Date"
            }
          },
          {
            "in": "query",
            "name": "to_date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Date"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "visit_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Visit Type"
            }
          },
          {
            "in": "query",
            "name": "location_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location Id"
            }
          },
          {
            "description": "Page number (1-based)",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number (1-based)",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Items per page",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "Items per page",
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Visits Api V1 Opd Visits Get",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Visits",
        "tags": [
          "OPD"
        ]
      },
      "post": {
        "description": "Create OPD visit (walk-in or appointment-linked).",
        "operationId": "create_visit_api_v1_opd_visits_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdVisitCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Visit Api V1 Opd Visits Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Visit",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/visits/{visit_id}": {
      "get": {
        "description": "Get OPD visit detail.",
        "operationId": "get_visit_api_v1_opd_visits__visit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Visit Api V1 Opd Visits  Visit Id  Get",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Visit",
        "tags": [
          "OPD"
        ]
      },
      "patch": {
        "description": "Update visit metadata before consultation.",
        "operationId": "update_visit_api_v1_opd_visits__visit_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdVisitUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Update Visit Api V1 Opd Visits  Visit Id  Patch",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Visit",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/visits/{visit_id}/cancel": {
      "post": {
        "description": "Cancel OPD visit.",
        "operationId": "cancel_visit_api_v1_opd_visits__visit_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpdVisitCancelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cancel Visit Api V1 Opd Visits  Visit Id  Cancel Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Visit",
        "tags": [
          "OPD"
        ]
      }
    },
    "/api/v1/opd/visits/{visit_id}/complete": {
      "post": {
        "description": "Complete consultation and finalize visit.",
        "operationId": "complete_visit_api_v1_opd_visits__visit_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OpdVisitCompleteRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Complete Visit Api V1 Opd Visits  Visit Id  Complete Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Visit",
        "tags": [
          "OPD"
        ]
      }
    },
...
    "/api/v1/opd/visits/{visit_id}/start": {
      "post": {
        "description": "Start doctor consultation.",
        "operationId": "start_visit_api_v1_opd_visits__visit_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "visit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Visit Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OpdVisitStartRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Start Visit Api V1 Opd Visits  Visit Id  Start Post",
                  "type": "null"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Visit",
        "tags": [
          "OPD"
        ]
      }
    },
...
  }
}